Reworked settings and run configuration

+++ src/routines/initialization/sobol.rs
@@ -8,7 +8,7 @@ use sobol_burley::sample;
 pub fn generate(
     n_points: usize,
     range_params: &Vec<(f64, f64)>,
-    seed: u32,
+    seed: usize,
 ) -> ArrayBase<OwnedRepr<f64>, Dim<[usize; 2]>> {
     let n_params = range_params.len();
     let mut seq = Array::<f64, _>::zeros((n_points, n_params).f());
@@ -16,7 +16,7 @@ pub fn generate(
         let mut row = seq.slice_mut(s![i, ..]);
         let mut point: Vec<f64> = Vec::new();
         for j in 0..n_params {
-            point.push(sample(i.try_into().unwrap(), j.try_into().unwrap(), seed) as f64)
+            point.push(sample(i.try_into().unwrap(), j.try_into().unwrap(), seed as u32) as f64)
         }
         row.assign(&Array::from(point));
     }
